requirements.txt
opencv-python==4.1.1.26
lxml
tqdm
tensorflow==2.3.0
absl-py
easydict
matplotlib
pillow

